+++ packages/sudoku-game/src/lib/Game/Controller/index.svelte
@@ -19,12 +19,11 @@
   } from '$lib/sudokuStore';
   import { scanner } from '$lib/sudokuStore/scanner';
   import type { InputMode, WalkthroughStep } from '@octopuzzles/models';
-  import { SquareButton, ControllerSkeleton, NotificationModal } from '@octopuzzles/ui';
+  import { SquareButton, ControllerSkeleton } from '@octopuzzles/ui';
   import { deepCopy, isCommandKey } from '@octopuzzles/utils';
   import ArrowCounterClockwise from 'phosphor-svelte/lib/ArrowCounterClockwise/ArrowCounterClockwise.svelte';
   import ArrowUUpLeft from 'phosphor-svelte/lib/ArrowUUpLeft/ArrowUUpLeft.svelte';
   import ArrowUUpRight from 'phosphor-svelte/lib/ArrowUUpRight/ArrowUUpRight.svelte';
-  import Check from 'phosphor-svelte/lib/Check/Check.svelte';
   import PersonSimpleWalk from 'phosphor-svelte/lib/PersonSimpleWalk/PersonSimpleWalk.svelte';
   import { onMount } from 'svelte';
   import { get } from 'svelte/store';
@@ -188,14 +187,6 @@
 
   const canUndo = gameHistory.canUndo;
   const canRedo = gameHistory.canRedo;
-
-  function verify(): void {
-    const errorCells = scanner.getErrorCells();
-    $wrongCells = errorCells;
-    constraintsChecked = errorCells.length === 0;
-  }
-
-  let constraintsChecked = false;
 </script>
 
 <svelte:window use:gameAction={{ onKeyDown: handleKeyboardShortcuts, onKeyUp: handleKeyUp }} />
@@ -241,14 +232,6 @@
   </svelte:fragment>
 
   <svelte:fragment slot="aux">
-    <button
-      title="Check digits"
-      class="w-8 h-8 hover:ring hover:ring-orange-500 rounded-full"
-      on:click={verify}
-    >
-      <Check size={32} />
-    </button>
-
     {#if $page.url.pathname.includes('/sudoku/editor') || walkthrough.length > 0}
       <button
         title="Walkthrough"
@@ -276,9 +259,4 @@
   bind:isOpen={walkthroughViewerModalIsOpen}
   clues={$clues}
   {walkthrough}
-/>
-
-<NotificationModal
-  bind:isOpen={constraintsChecked}
-  notificationMessage="No constraint violations detected"
 />
